removed createOktaAuthCoreLite
